Start refactoring of PTG classes
