+++ libs/nav/include/mrpt/nav/tpspace/CPTG1.h
@@ -6,10 +6,9 @@
    | See: http://www.mrpt.org/Authors - All rights reserved.                   |
    | Released under BSD License. See details in http://www.mrpt.org/License    |
    +---------------------------------------------------------------------------+ */
-#ifndef CPTG1_H
-#define CPTG1_H
+#pragma once
 
-#include <mrpt/nav/tpspace/CParameterizedTrajectoryGenerator.h>
+#include <mrpt/nav/tpspace/CPTG_DiffDrive_CollisionGridBased.h>
 
 namespace mrpt
 {
@@ -40,30 +39,23 @@ namespace mrpt
 	 *
 	 *  \ingroup nav_tpspace
 	 */
-	class NAV_IMPEXP CPTG1 : public CParameterizedTrajectoryGenerator
+	class NAV_IMPEXP CPTG1 : public CPTG_DiffDrive_CollisionGridBased
 	{
 	 public:
-			/** Constructor: possible values in "params", those of CParameterizedTrajectoryGenerator plus:
-			 *   - K: Direction, +1 or -1
-			 */
-			CPTG1(const mrpt::utils::TParameters<double> &params );
+		/** Constructor: possible values in "params", those of CParameterizedTrajectoryGenerator plus:
+			*   - K: Direction, +1 or -1
+			*/
+		CPTG1(const mrpt::utils::TParameters<double> &params );
 
-			virtual bool inverseMap_WS2TP(float x, float y, int &out_k, float &out_d, float tolerance_dist = 0.10f) const;
-
-			/** Gets a short textual description of the PTG and its parameters. */
-			std::string getDescription() const;
-
-
-			bool PTG_IsIntoDomain( float x, float y );
-			void PTG_Generator( float alpha, float t,float x, float y, float phi, float &v, float &w );
+		std::string getDescription() const MRPT_OVERRIDE;
+		bool inverseMap_WS2TP(double x, double y, int &out_k, double &out_d, double tolerance_dist = 0.10) const MRPT_OVERRIDE;
+		bool PTG_IsIntoDomain( double x, double y ) const MRPT_OVERRIDE;
+		void ptgDiffDriveSteeringFunction( float alpha, float t,float x, float y, float phi, float &v, float &w ) const MRPT_OVERRIDE;
 
 	 protected:
-			/** A generation parameter */
-			float	K;
+		/** A generation parameter */
+		double K;
 	};
   }
 }
-
-
-#endif
