+++ libs/nav/include/mrpt/nav/tpspace/CPTG2.h
@@ -6,10 +6,9 @@
    | See: http://www.mrpt.org/Authors - All rights reserved.                   |
    | Released under BSD License. See details in http://www.mrpt.org/License    |
    +---------------------------------------------------------------------------+ */
-#ifndef CPTG2_H
-#define CPTG2_H
+#pragma once
 
-#include <mrpt/nav/tpspace/CParameterizedTrajectoryGenerator.h>
+#include <mrpt/nav/tpspace/CPTG_DiffDrive_CollisionGridBased.h>
 
 namespace mrpt
 {
@@ -35,30 +34,21 @@ namespace mrpt
 	 *
 	 *  \ingroup nav_tpspace
 	 */
-	class NAV_IMPEXP  CPTG2 : public CParameterizedTrajectoryGenerator
+	class NAV_IMPEXP  CPTG2 : public CPTG_DiffDrive_CollisionGridBased
 	{
 	 public:
-			/** Constructor: possible values in "params", those of CParameterizedTrajectoryGenerator plus:
-			 *   - cte_a0v, cte_a0w: Parameters of this PTG (both are angles in radians).
-			 */
-			CPTG2(const mrpt::utils::TParameters<double> &params );
-
-			/** Gets a short textual description of the PTG and its parameters.
+		/** Constructor: possible values in "params", those of CParameterizedTrajectoryGenerator plus:
+			*   - cte_a0v, cte_a0w: Parameters of this PTG (both are angles in radians).
 			*/
-			std::string getDescription() const;
+		CPTG2(const mrpt::utils::TParameters<double> &params );
 
-			bool PTG_IsIntoDomain( float x, float y );
+		std::string getDescription() const MRPT_OVERRIDE;
+		bool inverseMap_WS2TP(double x, double y, int &out_k, double &out_d, double tolerance_dist = 0.10) const MRPT_OVERRIDE;
+		void ptgDiffDriveSteeringFunction( float alpha, float t,float x, float y, float phi, float &v, float &w ) const MRPT_OVERRIDE;
 
-			void PTG_Generator( float alpha, float t,float x, float y, float phi, float &v, float &w );
 	 protected:
-			float	cte_a0v;
-			float	cte_a0w;
-
-
+		double cte_a0v, cte_a0w;
 	};
   }
 }
-
-
-#endif
