+++ libs/nav/src/tpspace/CPTG6.cpp
@@ -1,4 +1,4 @@
-/* +---------------------------------------------------------------------------+
++---------------------------------------------------------------------------+
    |                     Mobile Robot Programming Toolkit (MRPT)               |
    |                          http://www.mrpt.org/                             |
    |                                                                           |
@@ -16,16 +16,10 @@ using namespace mrpt::nav;
 using namespace mrpt::system;
 using namespace mrpt::utils;
 
-/*---------------------------------------------------------------
-						Constructor
-  ---------------------------------------------------------------*/
-CPTG6::CPTG6(const mrpt::utils::TParameters<double> &params ) : CParameterizedTrajectoryGenerator(params)
+CPTG6::CPTG6(const mrpt::utils::TParameters<double> &params ) : CPTG_DiffDrive_CollisionGridBased(params)
 {
 }
 
-/*---------------------------------------------------------------
-						getDescription
-  ---------------------------------------------------------------*/
 std::string CPTG6::getDescription() const
 {
 	char str[100];
@@ -33,30 +27,24 @@ std::string CPTG6::getDescription() const
 	return std::string(str);
 }
 
-/*---------------------------------------------------------------
-						PTG_Generator
-  ---------------------------------------------------------------*/
-void CPTG6::PTG_Generator( float alpha, float t,float x, float y, float phi, float &v, float &w )
+void CPTG6::ptgDiffDriveSteeringFunction( float alpha, float t,float x, float y, float phi, float &v, float &w ) const
 {
 	MRPT_UNUSED_PARAM(t); MRPT_UNUSED_PARAM(x); MRPT_UNUSED_PARAM(y);
-    float At_a = alpha - phi;
+	float At_a = alpha - phi;
 
-    while (At_a>M_PI) At_a -= (float) M_2PI;
-    while (At_a<-M_PI) At_a += (float) M_2PI;
+	while (At_a>M_PI) At_a -= (float) M_2PI;
+	while (At_a<-M_PI) At_a += (float) M_2PI;
 
-    v = V_MAX * (1.0f - 0.95f * square( At_a / (float)M_PI ) );
-    if (At_a>0.01f)
-	    w = W_MAX;
-    else
-    	if (At_a<0.01f)
-             w = -W_MAX;
-        else w = 0;
+	v = V_MAX * (1.0f - 0.95f * square( At_a / (float)M_PI ) );
+	if (At_a>0.01f)
+		w = W_MAX;
+	else
+		if (At_a<0.01f)
+				w = -W_MAX;
+		else w = 0;
 }
 
-/*---------------------------------------------------------------
-					PTG_IsIntoDomain
-  ---------------------------------------------------------------*/
-bool CPTG6::PTG_IsIntoDomain( float x, float y )
+bool CPTG6::PTG_IsIntoDomain( double x, double y ) const
 {
 	MRPT_UNUSED_PARAM(x); MRPT_UNUSED_PARAM(y);
 	return true;
